# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Error completing request"
...
    SignUpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "password123"
    SignInRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "password123"
        guest_id:
          type: string
          example: "1"
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
    PasswordUpdateRequest:
      type: object
      properties:
        new_password:
          type: string
          example: "newpassword123"
...
    UserResponse:
      type: object
      properties:
        message:
          id: number
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        address:
          type: string
        email:
          type: string
        active:
          type: boolean
...
paths:
  /sign-up:
    post:
      tags:
        - Auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
      responses:
        "201":
          description: User successfully registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User successfully registered"
        "400":
          description: Email and password are required or invalid email format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to register user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sign-in:
    post:
      tags:
        - Auth
      summary: Sign in a user and return access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
      responses:
        "200":
          description: User successfully signed in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User successfully signed in"
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserResponse"
        "401":
          description: Invalid credentials or email does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Email and password are required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /token/refresh:
    post:
      tags:
        - Auth
      summary: Refresh the access token using a refresh token
      security:
        - BearerAuth: [ ]
      responses:
        "200":
          description: Access token successfully refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
  /reset-password:
    post:
      tags:
        - Auth
      summary: Request a password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: Password reset link has been sent to your email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset link has been sent to your email"
        "400":
          description: Email is required or invalid email format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /update-password/{reset_token}:
    post:
      tags:
        - Auth
      summary: Update password using reset token
      parameters:
        - in: path
          name: reset_token
          required: true
          schema:
            type: string
          description: The reset token received in the email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordUpdateRequest"
      responses:
        "200":
          description: Password has been updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password has been updated successfully"
        "400":
          description: New password is required, invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description:
...